:truck: :speech_balloon: Move `MainContent` to `app/components/MainContent`
